refactor(config): move .env/keyring writing to config/env_file.py
The generic '.env' writer lived in surfaces/cli/wizard/env_sync.py, so
only the onboarding wizard could reach it. 'opensre integrations setup'
(integrations/) and the interactive-shell action tools (tools/) cannot
import a surfaces module, which is why they save credentials to the
integration store only - and why an integration configured through
either one is invisible to the deploy preflight, which reads the env var
(platform/deployment/ecr_deploy/prep.py).

Move the generic half down to config/, the layer floor every tier may
import: sensitive-key classification, .env line rewriting, and the
keyring/'.env' routing (sync_env_secret / sync_env_values). What stays
in env_sync.py is the wizard's LLM-provider logic - which env keys a
provider owns and stripping the previous provider's keys on a switch.
No forwarding shim; the 17 importers move to the canonical path.

Behavior is unchanged. Both modules resolve the same PROJECT_ENV_PATH
via config.local_env.get_project_env_path().

sync_reasoning_model_env now resolves its default env path explicitly:
letting sync_env_values fall back to its own module-level default made
the two writers disagree about the target, which under test isolation
(which patches env_sync.PROJECT_ENV_PATH) escaped tmp_path and wrote to
the real project .env. Both .env-isolating test fixtures now pin the
config.env_file default too, so no test can reach the real file.
