+++ surfaces/cli/wizard/configurators/dagster.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+from config.env_file import sync_env_secret, sync_env_values
 from integrations.store import upsert_integration
 from platform.terminal.theme import SECONDARY
 from surfaces.cli.wizard._ui import (
@@ -11,7 +12,6 @@ from surfaces.cli.wizard._ui import (
     _render_integration_result,
     _string_value,
 )
-from surfaces.cli.wizard.env_sync import sync_env_secret, sync_env_values
 from surfaces.cli.wizard.integration_health import validate_dagster_integration
