+++ surfaces/cli/wizard/configurators/posthog.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from config.constants.posthog import DEFAULT_POSTHOG_URL
+from config.env_file import sync_env_secret, sync_env_values
 from integrations.store import upsert_integration
 from platform.terminal.theme import HIGHLIGHT, SECONDARY
 from surfaces.cli.wizard._ui import (
@@ -13,7 +14,6 @@ from surfaces.cli.wizard._ui import (
     _render_integration_result,
     _string_value,
 )
-from surfaces.cli.wizard.env_sync import sync_env_secret, sync_env_values
 from surfaces.cli.wizard.integration_health import (
     validate_posthog_integration,
     validate_posthog_mcp_integration,
