+++ tests/cli/test_model_persistence_scenarios.py
@@ -27,6 +27,7 @@ def persistence_paths(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> dict[s
     env_path = tmp_path / "project.env"
     store_path = tmp_path / "opensre.json"
     monkeypatch.setattr(env_sync, "PROJECT_ENV_PATH", env_path)
+    monkeypatch.setattr("config.env_file.PROJECT_ENV_PATH", env_path)
     monkeypatch.setattr(wizard_store, "get_store_path", lambda: store_path)
     return {"env": env_path, "store": store_path}
